Fix abi decoding of bytes32
Note that any indexed string or array will be stored as a hash, so
we need to make that clear. The new Hashed field marks that.

We might want to store strings in bytes32 so that it does not get
hashed. If this is the case then use code similar to the test
TestEventEmitterBytes32isString.

Signed-off-by: Sean Young <sean.young@monax.io>