Save binary at address when deploying bin file
When deploying a bin file, ensure that a copy is saved at the contract
address. This already happened when deploying a solidity file.

This makes it easier to call a function using burrow deploy, since the
abi does not need to specified with the bin parameter.

Refactor code to simplify.

Signed-off-by: Sean Young <sean.young@monax.io>