Fix restore bug
When loading a dump during restore, the events are restored without their
transactions. The TxHash field is all 0s; however we added a txHash
reference, which also has the b prefix. Then when iterating over the blocks,
we find the TxHash reference which is then decoded as a block, which fails.

Signed-off-by: Sean Young <sean.young@monax.io>