Decouple execution from Blockchain.
After moving validators into state we have an opportunity to increase
the decoupling between executor and Blockchain. This PR makes that move
by putting Blockchain committing back in the hands of the ABCI app and
getting all of the execution contexts to only ask for exactly what they
need from blockchain.

Signed-off-by: Silas Davis <silas@monax.io>