Reimplement query grammar to support OR
And also boolean logic with nesting and precedence.

The matching interface is cleaner and _should_ also be faster.

Stopped erasing type information for tagged objects by converting
everything to strings (only having to parse it back later).

With this grammar and interpreter it would be easy to add other helpers and operations.

Signed-off-by: Silas Davis <silas@monax.io>