Augment EVM Contract bytecode with Code object including OpcodeBitset for JUMPDEST analysis
The previous implmentation stored the bitset on the account, this had
the advantage of not having to recompute it for the same account, but
also complicated and made a breaking change to state. It also did not
consider the case when we are account creation initiailisation code
before the account has initialised. By using the Code type in
evm.Contract the type system helps us ensure the necessary OpcodeBitset
is initialised.

Signed-off-by: Silas Davis <silas@monax.io>