Implement cross-engine call dispatch for eWASM
- Implement the eWASM `call` syscall to support calling contracts

- Wire up WASM VM as a Disptacher so it can call EVM and Natives
(including Precompiles) as well as other WASM contracts.

- To support 128-bit wide (and 256-bit wide EVM!) values use big.Int for
Value and Gas. Note the Account model currently still uses uint64 so
outside of the execution machinery we still throw an error if a balance
would overflow uint64. This paves the way for relaxing this requirement
and making account balances etc.

- exec.CallData `Value` and `Gas` are now the big-endian big.Int bytes
rather than uint64 to support the above.

- EVM and WASM now share calling and some other code. EVM now more
strictly throws errors that are not ExecutionReverted becuase it
probably always should have and makes it more consistent with WASM.

WASM uses panics to transmit its errors. If we have errors that should
not halt execution (AKA 'non-trapping') we should probably explicitly encode
this fact into the coded errors and act accordingly and consistenly in
WASM and EVM

Signed-off-by: Silas Davis <silas@monax.io>